Snapfuzz: Check child_pid before group pid
There is a very elusive situation at which AFL might check if the group
pid is still alive before the target manages to register a group pid.

+++ afl-fuzz.c
@@ -1006,7 +1006,11 @@ typedef enum {
 } TargetAction;
 
 static boolean target_is_dead() {
-  int status = kill(-child_pid, 0);
+  int status = kill(child_pid, 0);
+  if (status == 0)
+    return FALSE;
+
+  status = kill(-child_pid, 0);
   if (status != 0) {
     if (errno == ESRCH) {
       return TRUE;
